Refactor: Moved the IO wrapping stuff to its own module (#1)
